+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGQ2CJCRFGMFMB9WW0FWH8WN"
+        "01KGQ98MQ8E5V3ZD7H84RTRYRR"
     )
     r"""The unique identifier of the data source"""