+++ packages/orq-rc/src/orq_ai_sdk/models/generateconversationnameop.py
@@ -186,7 +186,7 @@ class GenerateConversationNameResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kgq2chvhysx8n94hxz4s5srw"
+        "conv_01kgq98m3q6waanenyy1nprteh"
     )
     r"""Unique conversation identifier with `conv_` prefix."""